server/requirements.txt
langgraph
langgraph.prebuilt
langchain_openai
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
trustcall
langgraph-cli
python-dotenv
colorama
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
xgboost
scikit-learn
pandas
numpy